requirements.txt
Django==5.1.1
djangorestframework==3.15.2
pytest==8.3.3
django-admin==2.0.2
django-extensions==3.2.3
ipython==8.27.0
django-filter

